+++ product_reference/tests/__init__.py
@@ -2,3 +2,4 @@
 
 from . import test_base_reference
 from . import test_reference_uppercase
+from . import test_view_mounting